// packages/api/zendesk/infrastructure/constants/environment.ts
export enum ZendeskStackEnvironmentKeys {
  DD_VERSION = 'DD_VERSION',
  DD_ENV = 'DD_ENV',
  DD_API_KEY = 'DD_API_KEY',
  DD_GIT_COMMIT_SHA = 'DD_GIT_COMMIT_SHA',
  DD_GIT_REPOSITORY_URL = 'DD_GIT_REPOSITORY_URL',
  USE_DATADOG_AGENT = 'USE_DATADOG_AGENT',
  SKIP_ZENDESK_STACK = 'SKIP_ZENDESK_STACK',
  ZENDESK_APP_CLIENT_ID = 'ZENDESK_APP_CLIENT_ID',
  ZENDESK_REDIRECT_URI = 'ZENDESK_REDIRECT_URI',
  USER_POOL_DOMAIN = 'USER_POOL_DOMAIN',
  ZENDESK_JWT_SECRET = 'ZENDESK_JWT_SECRET',
  ZENDESK_MESSAGING_JWT_SECRET = 'ZENDESK_MESSAGING_JWT_SECRET',
  ZENDESK_MESSAGING_KID = 'ZENDESK_MESSAGING_KID',
  ZENDESK_SUBDOMAIN = 'ZENDESK_SUBDOMAIN',
  DEBUG_LOGGING_ENABLED = 'DEBUG_LOGGING_ENABLED',
  SERVICE = 'SERVICE',
  ZENDESK_URL_UK = 'ZENDESK_URL_UK',
  ZENDESK_URL_AUS = 'ZENDESK_URL_AUS',
  ZENDESK_URL_DDS = 'ZENDESK_URL_DDS',
  ZENDESK_SUPPORT_URL_UK = 'ZENDESK_SUPPORT_URL_UK',
  ZENDESK_SUPPORT_URL_AUS = 'ZENDESK_SUPPORT_URL_AUS',
  ZENDESK_SUPPORT_URL_DDS = 'ZENDESK_SUPPORT_URL_DDS',
  ZENDESK_API_BASE_URL_UK = 'ZENDESK_API_BASE_URL_UK',
  ZENDESK_API_BASE_URL_AUS = 'ZENDESK_API_BASE_URL_AUS',
  ZENDESK_API_BASE_URL_DDS = 'ZENDESK_API_BASE_URL_DDS',
}